refactor: separate remained 4 graphs
